fix(a11y): move focus back to the container
